fixing redirect page, login, logout

+++ rms-servlet-web/src/main/java/com/mitrais/rms/controller/UserServlet.java
@@ -9,6 +9,8 @@ import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
+
 import java.io.IOException;
 import java.util.List;
 import java.util.Optional;
@@ -20,25 +22,34 @@ public class UserServlet extends AbstractController
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException
     {
-        String path = getTemplatePath(req.getServletPath()+req.getPathInfo());
+    	String path = getTemplatePath(req.getServletPath()+req.getPathInfo());
+    	HttpSession session = req.getSession(true);
+    	if(session.getAttribute("username") == null)
+		{
+    		resp.sendRedirect(req.getContextPath()+"/login");
+		}
+		else {
+			
 
-        if ("/list".equalsIgnoreCase(req.getPathInfo())){
-            List<User> users = userDao.findAll();
-            req.setAttribute("users", users);
-        }
-        if ("/edit".equalsIgnoreCase(req.getPathInfo())){
-        	Long id = Long.parseLong(req.getParameter("id"));
-            User user = userDao.find(id);
-            req.setAttribute("user", user);
-        }
-        if("/delete".equalsIgnoreCase(req.getPathInfo())){
-        	Long id = Long.parseLong(req.getParameter("id"));
-        	User user = userDao.find(id);
-            userDao.delete(user);
-            path = null;
-        }
-        RequestDispatcher requestDispatcher = req.getRequestDispatcher(path);
-        requestDispatcher.forward(req, resp);
+	        if ("/list".equalsIgnoreCase(req.getPathInfo())){
+	            List<User> users = userDao.findAll();
+	            req.setAttribute("users", users);
+	        }
+	        if ("/edit".equalsIgnoreCase(req.getPathInfo())){
+	        	Long id = Long.parseLong(req.getParameter("id"));
+	            User user = userDao.find(id);
+	            req.setAttribute("user", user);
+	        }
+	        if("/delete".equalsIgnoreCase(req.getPathInfo())){
+	        	Long id = Long.parseLong(req.getParameter("id"));
+	        	User user = userDao.find(id);
+	            userDao.delete(user);
+	            path = null;
+	        }
+	        RequestDispatcher requestDispatcher = req.getRequestDispatcher(path);
+	        requestDispatcher.forward(req, resp);
+		}
+        
     }
     
     @Override
